Refactor: Create apis/useContentEditMutation
